src/requirements.txt
dnspython3==1.12.0
secure-smtpd==3.0.0
